deleted apply folder
